remove JsonIO, welcome back JsonIOWrapper
